UGENE-3262 fix: GUI tests
git-svn-id: https://local.ugene.unipro.ru/svn/ugene/trunk@7305 7ca3a678-163c-e011-b84f-001e8c0267ff

+++ src/corelibs/U2Formats/src/GenbankPlainTextFormat.cpp
@@ -854,7 +854,7 @@ static void prepareMultiline(QString& line, int spacesOnLineStart, bool lineBrea
                         skipLineBreak = true;
                     }
                 }
-                newLine.append(line.mid(pos, pos2 - pos + (int)(!lineBreakOnlyOnSpace || skipLineBreak)));
+                newLine.append(line.mid(pos, pos2 - pos + (!lineBreakOnlyOnSpace || skipLineBreak)));
             } else { //last line
                 newLine.append(line.mid(pos, len - pos));
             }
